Fixed compiler warnings.
git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/main/trunk@8327 178a84e3-b1eb-0310-8ba1-8eac791a3b58

+++ src/input/input_manager.cpp
@@ -624,9 +624,9 @@ EventPropagation InputManager::input(const SEvent& event)
         {
             // *0.017453925f is to convert degrees to radians
             dispatchInput(Input::IT_STICKMOTION, event.JoystickEvent.Joystick, Input::HAT_H_ID, Input::AD_NEUTRAL,
-                          cos(event.JoystickEvent.POV*0.017453925f/100.0f)*Input::MAX_VALUE);
+                          (int)(cos(event.JoystickEvent.POV*0.017453925f/100.0f)*Input::MAX_VALUE));
             dispatchInput(Input::IT_STICKMOTION, event.JoystickEvent.Joystick, Input::HAT_V_ID, Input::AD_NEUTRAL,
-                          sin(event.JoystickEvent.POV*0.017453925f/100.0f)*Input::MAX_VALUE);
+                          (int)(sin(event.JoystickEvent.POV*0.017453925f/100.0f)*Input::MAX_VALUE));
         }
         
         GamePadDevice* gp = getDeviceList()->getGamePadFromIrrID(event.JoystickEvent.Joystick);
